VAEs/requirements.txt
torch==1.8.0
torchvision==0.9.0
scipy==1.4.1
tqdm==4.50.2
adacvar==0.0.1
matplotlib==3.2.2
numpy==1.18.4
networkx==2.5
Pillow==8.1.2
scikit_learn==0.24.1
